1. SequenceSet.cpp: typo editing 2. Alphabet -> add debug() for exhaustive print out 3. Global / main -> call debug-functions from within main instead of in Global 4. Global -> use size_t instead of int in for-loops going over vector until vector.size (instead of casting vector.size to int) 5. Sequence / SequenceSet: remove friend-class classifications

+++ src/shared/SequenceSet.cpp
@@ -29,8 +29,8 @@ SequenceSet::SequenceSet( std::string sequenceFilepath, bool revcomp, std::strin
 
 SequenceSet::~SequenceSet(){
 	if( baseFrequencies_ ) delete[] baseFrequencies_;
-	for (int i = 0; i < sequences_.size(); ++i) {
-	    delete sequences_[i]; // Calls ~object (which deallocates sequences_[i]->sequence_)and deallocates *sequences_[i]
+	for (size_t i = 0; i < sequences_.size(); ++i) {
+	    delete sequences_[i]; // Calls ~sequence_ (which deallocates sequences_[i]->sequence_)and deallocates *sequences_[i]
 	}
 }
